Adjust FileContentElement tests
